erro logica

+++ src/pt/ulusofona/lp2/deisiGreatGame/GameManager.java
@@ -337,7 +337,7 @@ public class GameManager {
                         } // erro de sintaxe - feito
                         case 1 -> {
                             if (!tools.get(2).contains(jogada.get(0)) || !tools.get(5).contains(jogada.get(0))) {
-                                double dado = (double) dados / 2;
+                                int dado = dados / 2;
                                 programmers.get(jogada.get(0)).setPos(posicaoAtual - (int) Math.floor(dado));
                                 explicacao = "Recua " + dado + " casa(s)";
                             } else {
